panic() if we ever get disconnected from Firestore for watching transactions or entities
We can't safely recover from these scenarios at the moment, because even if we reconnect, there's a chance the incoming messages don't represent a reconstructed history. And because at this point, configstore isn't running in a "reconstruct initial state" mode, it's possible for it to get stalled out waiting for transaction components that never arrive.

It's better to panic() here, let configstore restart and do the proper initial state reconstruction before propagating the whole new state down to clients.